services/whisper_transcriber/requirements.txt
# Requirements placeholder

fastapi==0.104.1
uvicorn==0.24.0
torch==2.1.1
numpy==1.24.3
openai-whisper==20231117
ffmpeg-python==0.2.0
python-multipart==0.0.6
pydantic==2.5.2
python-dotenv==1.0.0
